Update mouse position using a callback + misc stuff
