+++ CMakeLists.txt
@@ -28,7 +28,7 @@ elseif ( "${CMAKE_CXX_COMPILER_ID}" MATCHES "Clang" )
         -Wimplicit-fallthrough
         -Wno-missing-field-initializers
     )
-else ( "${CMAKE_CXX_COMPILER_ID}" MATCHES "GNU" )
+elseif ( "${CMAKE_CXX_COMPILER_ID}" MATCHES "GNU" )
     add_compile_options (
         -Wall
         -Wextra
@@ -51,6 +51,8 @@ else ( "${CMAKE_CXX_COMPILER_ID}" MATCHES "GNU" )
         -Wimplicit-fallthrough
         -Wno-missing-field-initializers
     )
+else ()
+    message(FATAL_ERROR "Unsupported compiler: [${CMAKE_CXX_COMPILER_ID}]")
 endif ()
 
 option ( LINT OFF )
