+++ include/glfw/glfw/input/mouse.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "glfw/core.h"
+#include "glfw/input.h"
 
 namespace glfw::input
 {
@@ -11,7 +12,7 @@ struct MouseMovement
 {
     double dx = 0.0;
     double dy = 0.0;
-};
+}; // MouseMovement
 
 struct MousePosition
 {
@@ -23,7 +24,7 @@ struct MousePosition
     {
         return MouseMovement{ lhs.x - rhs.x, lhs.y - rhs.y };
     }
-};
+}; // MousePosition
 
 class MouseHandler
 {
@@ -33,6 +34,12 @@ class MouseHandler
         instance( window ).buttonCallback( key, static_cast<ButtonAction>( action ), ModifierFlag{ mods } );
     }
 
+    static void positionCallbackWrapper( GLFWwindow* window, double x, double y )
+    {
+        instance( window ).positionCallback( x, y );
+    }
+
+  private:
     void buttonCallback( MouseButtonIndex key, ButtonAction action, ModifierFlag modifier )
     {
         assert( action != ButtonAction::e_repeat );
@@ -41,17 +48,23 @@ class MouseHandler
             auto g = std::lock_guard{ m_mx };
             m_event_map[ key ].pushEvent( ButtonEvent{ .mods = modifier, .action = action } );
         }
-    }
+    } // buttonCallback
+
+    void positionCallback( double x, double y )
+    {
+        auto g = std::lock_guard{ m_mx };
+        m_position = MousePosition{ .x = x, .y = y };
+    } // positionCallback
 
   private:
     MouseHandler( GLFWwindow* window )
         : m_window{ window }
     {
         glfwSetMouseButtonCallback( window, buttonCallbackWrapper );
-        m_position = getPosition();
+        glfwSetCursorPosCallback( window, positionCallbackWrapper );
     }
 
-    MousePosition updatePosition() { return std::exchange( m_position, getPosition() ); }
+    MousePosition updatePosition() { return std::exchange( m_old_position, m_position ); }
 
   public:
     using ButtonEventMap = std::unordered_map<MouseButtonIndex, ButtonEventInfo>;
@@ -60,19 +73,11 @@ class MouseHandler
         return *s_handler_table.lookup( window, [ window ]() {
             return std::unique_ptr<MouseHandler>{ new MouseHandler{ window } };
         } );
-    }
+    } // instance
 
   public:
     void setHidden() { glfwSetInputMode( m_window, GLFW_CURSOR, GLFW_CURSOR_HIDDEN ); }
     void setNormal() { glfwSetInputMode( m_window, GLFW_CURSOR, GLFW_CURSOR_NORMAL ); }
-
-  private:
-    MousePosition getPosition()
-    {
-        MousePosition ret;
-        glfwGetCursorPos( m_window, &ret.x, &ret.y );
-        return ret;
-    }
 
   public:
     struct PollResult
@@ -80,7 +85,7 @@ class MouseHandler
         ButtonEventMap buttons;
         MousePosition position;
         MouseMovement movement;
-    };
+    }; // PollResult
 
     PollResult poll()
     {
@@ -95,16 +100,17 @@ class MouseHandler
         auto movement = m_position - old_position;
 
         return PollResult{ std::move( old_map ), m_position, movement };
-    }
+    } // poll
 
   private:
     ButtonEventMap m_event_map;
-    MousePosition m_position;
+    MousePosition m_position;     // Current position
+    MousePosition m_old_position; // Position from the previous poll call
     GLFWwindow* const m_window;
     mutable std::mutex m_mx;
 
   private:
     static inline detail::GlobalHandlerTable<std::unique_ptr<MouseHandler>> s_handler_table;
-};
+}; // MouseHandler
 
 } // namespace glfw::input
